packages/remoting_mcp/requirements.txt
fastmcp>=0.4.1
livekit-agents>=1.0.0
livekit-api>=1.0.0
pynput>=1.7.7
mss>=9.0.1
pywin32>=306
python-dotenv>=1.0.0
